Refactor:
* What do LDCEs do?

* LDCEs allow to execute:

** Metadata queries

** Analytical queries

* Metadata queries allow parameters to filter for certain elements.

* The schema of results of metadata queries is inspired by olap4j result
sets.

* How can LDCEs defer?

** LDCEs may execute metadata queries differently. E.g., they may
directly be run on a pre-populated triple store. Or, they may partly
populate a triple store on-demand. Or, they may execute queries directly
on Linked Data sources.

**

* EmbeddedSesameEngine - Metadata/Analytical queries on an embedded
Sesame store that gets populated on-demand.

* OpenVirtuosoEngine - Metadata/Analytical queries on a pre-populated
Open Virtuoso store (used in SMART Data Explorer and FIOS)

* SesameEngine - Metadata/Analytical queries on a pre-populated Sesame
store (considered for use in ls3finance thesis)

* QcrumbEngine - Metadata/Analytical queries using Qcrumb.com
(considered for Slicer)

* We are interested in further extending the EmbeddedSesameEngine
(currently done in GDP_Embedded...)

** Should implement properly defined analytical operators (Logical
Operators, Physical Iterators)

** Should allow drill-across (having several datasets in FROM clause)

** Should allow to explicitly convert datasets, i.e., to refer to
derived datasets.

* Loading of data cubes (Look-ups, Normalisation, Reasoning, Integrity
checks) can be done in different ways:

** Coded in Java.

** Coded in Data-Fu. Then, we may also automatically derive new datasets
and define a global dataset.

* I am wondering how I should be implementing Base-cube, projection,
dice, slice, roll-up.

** Using a direct translation to one SPARQL query. <= This would be the
goal of LDCX (we do not consider multi-level hierarchies, but
drill-across) and follow-up work on olap4ld.

** Creating intermediary datasets using several SPARQL queries. <= This
would be the goal of the Macro-modelling paper (we consider multi-level
hierarchies, drill-across, conversion functions) and follow-up work with
Data-Fu, Dimension hierarchies, Slicer.

* Do we need separate EmbeddedSesameEngines?

Not necessarily, since both use an embedded Sesame engine, but we need
separate Physical Iterators. The EmbeddedSesameEngine always: 1) Takes
in and transforms a logical query plan into a physical query plan; 2)
Iterates through the root of the physical query plan returning Node[]
for each single result and creates List<Node[]>. The question is: Is the
header also returned? Does the header contain any meaning? Actually,
from a set of logical operators and metadata of input datasets, we can
automatically derive the metadata of the output data cube which would
align with the header/output of List<Node[]>. For now, we assume a
header is returned but not taken into account.

Connection allows to getMetaData() and to createStatement().
DatabaseMetadata allows to execute meta data queries. Statement allows
to execute analytical (MDX) queries and returns CellSet. CellSet allows
to ask for cells in a pivot table. CellSet creates an initial logical
query plan from a select statement (parsed from MDX syntax tree) and
executes the logical query plan on the selected Linked Data Cubes
engine. The result is List<Node[]> and is cached, the header is ignored.

The ordering of dimensions and measures in any element of List<Node[]>
is given by the list of hierarchies of each axis metadata of the pivot
table. This ordering is defined by the MDX query. How is this ordering
related to 1) the ordering of dimensions in the queried cubes and 2) the
logical query plan? Previously, the analytical query was issued with
"executeOlapQuery(Cube cube, List<Level> slicesrollups, List<Position>
dices, List<Measure> projections)" which would explicitly encode an
ordering. Now, the ordering is given by the ordering in the queried
cubes. Thus, we need to match the list of hierarchies in the cellset
with the list of hierarchies in the queried cubes.

Or, we build a cache that orders the dimensions in the ordering of the
queried cubes and that for any cell query first changes the ordering
according to the queried cube(s).

XXX do that

In any case, the header is ignored.

Another open issue is the "measures" dimension. Olap4j assumes that it
is existing but there are no direct correspondences in Linked Data (only
if we assume the measure property to be dimension, hierarchy, level,
member).

For now, we assume it is queried.

An LDCE takes in a logical query plan and runs getExecplan() to create a
physical query plan. For that it uses a
LogicalOlapOperatorQueryPlanVisitor.

What LogicalOlapOperatorQueryPlanVisitor do we have?

* Olap2SparqlSesameDerivedDatasetVisitor <= Was extended for
Drill-across but now is more going towards "deriving datasets". Maybe,
only for ConvertContext, we should materialise a new dataset.

* Olap2SparqlSesameVisitor <= Should also for LDCX eventually support
Drill-across but for now does not properly define logical operators.
Assumes that rollup is executed for all levels, even if not rolled-up.

To have several engines has the disadvantage that they share some
functionality/logic, e.g., metadata queries, loading of cubes.

If I see this correctly, we may for now have one engine that uses
different

* Analytical operators

* Drill-across operator

* Conversion operator (Convert-cube) <= see Split...