* Convert-context for one cube implemented. Now needs to be extended to any number of datasets.
* Question: How is the metadata managed?
** LogicalOlapOperator always have as input LogicalOlapOp(s) so that
they can be nested. Also, the have as input parameters that define what
the operator does, e.g., for projection, the projected measures are
given, for slice, the sliced dimensions are given. However, the metadata
of the cube is only implicit.
** PhysicalOlapIterator always have as input PhysicalOlapIterator(s) so
that they can be nested. For communication between iterators, a RDF
repository as temporary table is used. A PhysicalOlapIterator 1) takes
RDF from repo and inserts RDF for both data and metadata to repo 2)
returns observations with dimensions and measures and 3) allows to ask
for the metadata of the cube (using public attributes). Since a
PhysicalOlapIterator for the metadata can asks its input
PhysicalOlapIterators, we do not need to have as parameter the metadata,
but only the parameters for the operator.

* Tests - Problems
** Measures Dimension - Just ignored
** Implict Measures AGGFUN - Just ignored
** ? missing - MakeVariable refactored
** bindas! - Just ignored
** dsd? - Add same dsd to new ds.
** Function has to be properly: ((?boneg * 1000000) as ?boneg_eur)