* First working pipeline.
* We probably have to distinguish body and head
** For body, we fill each dataset with the dimensions of the first (and
not second cube) (that are not mentioned, already). For measures, we
fill each dataset with its own measures. 
** For head, we fill the dataset with the dimensions of the first cube
(that are not mentioned, already). For measures, we fill the dataset
with the measures of the first cube (that are not mentioned, already).


Note, in this case we assume that any conversion function is written so
that: 1) The dsd of the first dataset is retained. 2) All dimensions of
the second dataset are either fixed or also contained in first dataset.
3) if a second dataset is inserted, then a variable is given for that
dataset in the conversion function.

Problem: Iterators do not propagate their results. In theory, the name
of the dataset changes. I.e., we have to change the metadata. 

* Problem: What about the metadata of getCubes() etc.?
** getMap specifically was introduced so that no fixed ordering is
needed.
** Thus, to change the metadata, we 1) simply copy over all values. 2)
specifically change values as needed.
** Maybe, I should in fact fix the schema and the ordering?

Tests - Problems
* DSDs do not need to be the same for input datasets.
* Propagation of results do not seem to work. Maybe it is because of DSD
triples? For debugging, make RDF dump at RDF generation, querying. (find
http://estatwrap.ontologycentral.com/dsd/nama_gdp_c#dsd)
** dsd seems not to be given to new datasets. Therefore, the construct
query does not work. Instead, we add the dsd separately (adding
triples).
** Heapspace in Eclipse startup erhöht:
<pre>
-Xms512m
-Xmx1024m
-XX:PermSize=256M
-XX:MaxPermSize=512M
</pre>
* also in eclipse.ini <= but was not needed, actually.